+++ FSG/Entities/Modifier.cs
@@ -17,7 +17,7 @@ namespace FSG.Entities
 
         public ModifierType ModifierType { get; init; }
 
-        public int Value { get; init; }
+        public decimal Value { get; init; }
 
         public int RemainingTurns { get; set; }
